+++ packages/core-gatsby-ui/src/hooks/use-docs-navigation-data.hook.tsx
@@ -30,6 +30,22 @@ const query = graphql`
           locale
           layout
           template
+          frontmatter {
+            title
+            subject
+            tags
+            description
+            version
+            published
+            status
+            deprecated
+            editPageUrl
+            nextPageLabel
+            nextPageUrl
+            componentStatus
+            componentVersion
+            componentTests
+          }
         }
       }
     }
@@ -57,6 +73,22 @@ const query = graphql`
           locale
           layout
           template
+          frontmatter {
+            title
+            subject
+            tags
+            description
+            version
+            published
+            status
+            deprecated
+            editPageUrl
+            nextPageLabel
+            nextPageUrl
+            componentStatus
+            componentVersion
+            componentTests
+          }
         }
       }
     }