+++ packages/core-gatsby-ui/src/layout/docs.layout.tsx
@@ -1,5 +1,5 @@
 import { SITE_LANGUAGES } from '@newrade/core-common';
-import { HEADING, PARAGRAPH_SIZE, VIEWPORT } from '@newrade/core-design-system';
+import { HEADING, PARAGRAPH_SIZE, TagSize, Variant, VIEWPORT } from '@newrade/core-design-system';
 import { GatsbyMarkdownFilePageContext } from '@newrade/core-gatsby-config';
 import { SOURCE_INSTANCE_NAME } from '@newrade/core-gatsby-config/lib/esm/config/gatsby-source-instances';
 import { GatsbyLink, NavbarDocs } from '@newrade/core-gatsby-ui/src';
@@ -14,6 +14,7 @@ import {
   MainWrapper,
   SidebarItem,
   Stack,
+  Tag,
   useTreatTheme,
   useViewportBreakpoint,
   viewportContext,
@@ -122,13 +123,31 @@ export const LayoutDocs = React.memo<LayoutDocsProps>((props) => {
                         {item.items?.length ? (
                           <Stack>
                             {item.items?.map((item, itemIndex) => {
+                              const status = item.frontmatter?.status;
+                              const version = item.frontmatter?.version;
+                              const deprecated = item.frontmatter?.deprecated;
+
                               return (
                                 <DesktopDocsSidebarItem
                                   key={itemIndex}
                                   active={item.path === props.location?.pathname}
                                   AsElement={<GatsbyLink to={item.path} noStyles={true} />}
                                 >
-                                  {item.displayName || item.name}
+                                  <span style={{ marginRight: 4 }}>
+                                    {item.displayName || item.name}
+                                  </span>{' '}
+                                  {version ? (
+                                    <Tag
+                                      size={TagSize.small}
+                                      variant={Variant.tertiary}
+                                    >{`${version}`}</Tag>
+                                  ) : null}{' '}
+                                  {status ? (
+                                    <Tag
+                                      size={TagSize.small}
+                                      variant={Variant.secondary}
+                                    >{`${status.toUpperCase()}`}</Tag>
+                                  ) : null}{' '}
                                 </DesktopDocsSidebarItem>
                               );
                             })}