+++ packages/core-react-ui/src/tag/tag.tsx
@@ -1,30 +1,42 @@
-import { TagProps, Variant } from '@newrade/core-design-system';
+import { TagProps, TagSize, Variant } from '@newrade/core-design-system';
 import React from 'react';
 import { useStyles } from 'react-treat';
-import { useCommonProps } from '../hooks/use-common-props.hook';
+import { Primitive } from '../primitive/primitive';
 import { PrimitiveProps } from '../primitive/primitive.props';
-import { Label } from '../text/label';
 import * as stylesRef from './tag.treat';
 
-type Props = PrimitiveProps &
+type Props = PrimitiveProps<'div'> &
   TagProps & {
     children?: string;
   };
 
 export const Tag: React.FC<Props> = React.memo(
-  ({ id, style, className, children, as, variant, ...props }) => {
+  ({ id, style, className, children, as, variant, kind, size = TagSize.medium, ...props }) => {
     const styles = useStyles(stylesRef);
-    const commonProps = useCommonProps({
-      id,
-      style,
-      className,
-      classNames: [styles.base, variant ? styles[variant] : styles[Variant.primary]],
-    });
 
     return (
-      <div {...commonProps}>
-        <Label className={styles.label}>{children}</Label>
-      </div>
+      <Primitive
+        {...{
+          id,
+          style,
+          className,
+          classNames: [
+            styles.base,
+            size ? styles[size] : styles[TagSize.medium],
+            variant ? styles[variant] : styles[Variant.primary],
+          ],
+          ...props,
+        }}
+      >
+        <Primitive
+          classNames={[
+            styles.label,
+            size === TagSize.medium ? styles.mediumLabel : styles.smallLabel,
+          ]}
+        >
+          {children}
+        </Primitive>
+      </Primitive>
     );
   }
 );